Overview: Reworked Element.go. Details: 1. Found a temporary solution to the initialization problem by using the lowest bit to mark it as unused. 2. Improved BitArray.go. 3. Changed the signature of the export of runtime.memhash. TODO: 1. Optimize linked list variant more. 2. Add delete function potentially with optimize cache line. 3. Add the option to cache hash values for faster resize.

+++ Maps/Utils.go
@@ -49,4 +49,5 @@ func Max(a, b int) int {
 }
 
 //go:linkname RTHash runtime.memhash
-func RTHash(ptr unsafe.Pointer, seed, len uintptr) uint64
+//go:noescape
+func RTHash(ptr unsafe.Pointer, seed, len uintptr) uintptr
